Move grype vulnerability mapper logic and test to simplify testing.
